Move closing expired position into dedicated module
