+++ coordinator/src/bin/coordinator.rs
@@ -1,25 +1,20 @@
 use anyhow::Context;
 use anyhow::Result;
 use coordinator::cli::Opts;
-use coordinator::db;
 use coordinator::logger;
 use coordinator::metrics;
 use coordinator::metrics::init_meter;
 use coordinator::node;
 use coordinator::node::connection;
+use coordinator::node::expired_positions;
 use coordinator::node::storage::NodeStorage;
 use coordinator::node::Node;
-use coordinator::node::TradeAction;
-use coordinator::position::models::Position;
-use coordinator::position::models::PositionState;
 use coordinator::routes::router;
 use coordinator::run_migration;
 use coordinator::settings::Settings;
 use diesel::r2d2;
 use diesel::r2d2::ConnectionManager;
 use diesel::PgConnection;
-use hex::FromHex;
-use lightning::ln::PaymentHash;
 use lightning::util::events::Event;
 use ln_dlc_node::seed::Bip39Seed;
 use rand::thread_rng;
@@ -30,15 +25,13 @@ use std::net::Ipv4Addr;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use std::time::Duration;
-use time::OffsetDateTime;
 use tokio::sync::watch;
 use tokio::task::spawn_blocking;
 use tracing::metadata::LevelFilter;
-use trade::bitmex_client::BitmexClient;
 
 const PROCESS_PROMETHEUS_METRICS: Duration = Duration::from_secs(10);
 const PROCESS_INCOMING_DLC_MESSAGES_INTERVAL: Duration = Duration::from_secs(5);
-const POSITION_SYNC_INTERVAL: Duration = Duration::from_secs(300);
+const EXPIRED_POSITION_SYNC_INTERVAL: Duration = Duration::from_secs(300);
 const CONNECTION_CHECK_INTERVAL: Duration = Duration::from_secs(30);
 
 #[tokio::main]
@@ -171,107 +164,8 @@ async fn main() -> Result<()> {
         let node = node.clone();
         async move {
             loop {
-                tokio::time::sleep(POSITION_SYNC_INTERVAL).await;
-
-                let mut conn = match node.pool.get() {
-                    Ok(conn) => conn,
-                    Err(e) => {
-                        tracing::error!("Failed to get pool connection. Error: {e:?}");
-                        continue;
-                    }
-                };
-
-                let positions = match db::positions::Position::get_all_open_positions(&mut conn) {
-                    Ok(positions) => positions,
-                    Err(e) => {
-                        tracing::error!("Failed to get positions. Error: {e:?}");
-                        continue;
-                    }
-                };
-
-                let positions = positions
-                    .into_iter()
-                    .filter(|p| {
-                        p.position_state == PositionState::Open
-                            && OffsetDateTime::now_utc().ge(&p.expiry_timestamp)
-                    })
-                    .collect::<Vec<Position>>();
-
-                for position in positions.iter() {
-                    tracing::trace!(trader_pk=%position.trader, %position.expiry_timestamp, "Attempting to close expired position");
-
-                    if !node.is_connected(&position.trader) {
-                        tracing::debug!(
-                            "Could not close expired position with {} as trader is not connected.",
-                            position.trader
-                        );
-                        continue;
-                    }
-
-                    let channel_id = match node.decide_trade_action(&position.trader) {
-                        Ok(TradeAction::Close(channel_id)) => channel_id,
-                        Ok(_) => {
-                            tracing::error!(
-                                ?position,
-                                "Unable to find sub channel of expired position."
-                            );
-                            continue;
-                        }
-                        Err(e) => {
-                            tracing::error!(
-                                ?position,
-                                "Failed to decide trade action. Error: {e:?}"
-                            );
-                            continue;
-                        }
-                    };
-
-                    let closing_price =
-                        match BitmexClient::get_quote(&position.expiry_timestamp).await {
-                            Ok(quote) => match position.direction {
-                                trade::Direction::Long => quote.bid_price,
-                                trade::Direction::Short => quote.ask_price,
-                            },
-                            Err(e) => {
-                                tracing::warn!(
-                                    "Failed to get quote from bitmex for {} at {}. Error: {e:?}",
-                                    position.trader,
-                                    position.expiry_timestamp
-                                );
-                                continue;
-                            }
-                        };
-
-                    // Upon collab closing an expired position we cannot charge a fee using an
-                    // invoice. This dummy hash exists in the database to
-                    // represent zero-amount invoices.
-                    let zero_amount_payment_hash_dummy = PaymentHash(
-                        <[u8; 32]>::from_hex(
-                            "6f9b8c95c2ba7b1857b19f975372308161fedf50feb78a252200135a41875210",
-                        )
-                        .expect("static payment hash to decode"),
-                    );
-
-                    match node
-                        .close_position(
-                            position,
-                            closing_price,
-                            channel_id,
-                            zero_amount_payment_hash_dummy,
-                        )
-                        .await
-                    {
-                        Ok(_) => tracing::info!(
-                            "Successfully proposed to close expired position with {}",
-                            position.trader
-                        ),
-                        Err(e) => tracing::warn!(
-                            ?position,
-                            "Failed to close expired position with {}. Error: {e:?}",
-                            position.trader
-                        ),
-                    }
-                }
+                tokio::time::sleep(EXPIRED_POSITION_SYNC_INTERVAL).await;
+                expired_positions::close(node.clone()).await;
             }
         }
     });
